+++ src/main/java/com/openclassrooms/chatop/repositories/RentalRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 // Indique que cette interface est un dépôt Spring pour l'entité Rental
 @Repository
-public interface RentalRepository extends JpaRepository<Rental, Long> {
+public interface RentalRepository extends JpaRepository<Rental, Integer> {
 
     // Méthode pour trouver les locations par ownerId
-    List<Rental> findByOwnerId(Long ownerId);
+    List<Rental> findByOwnerId(Integer ownerId);
 }
